+++ packages/components/src/AtlantisPortalContent/AtlantisPortalContent.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { AtlantisPortalContentProps } from "./AtlantisPortalContent.types";
+import {
+  AtlantisThemeContextProvider,
+  useAtlantisTheme,
+} from "../AtlantisThemeContext";
+
+export function AtlantisPortalContent({
+  children,
+}: AtlantisPortalContentProps) {
+  const { theme } = useAtlantisTheme();
+
+  return (
+    <AtlantisThemeContextProvider dangerouslyOverrideTheme={theme}>
+      {children}
+    </AtlantisThemeContextProvider>
+  );
+}